Fixed a bug which occured when y was negative. I believe this was caused by comparing a signed integer to a floating point number. The compiler probabily was interpreting y as MAX_INT

+++ src/terminal/terminal_libx11.cpp
@@ -118,6 +118,7 @@ namespace LOG
 				}
 			}
 		}
+
 		update();
 	}
 
@@ -148,10 +149,10 @@ namespace LOG
 		if (!msgs.empty()) {
 
 			int y = 1 - linePos;
-			for (auto it = msgs.begin() + msgsPos; it != msgs.end() && y <= (height / 16); ++it) {
+			for (auto it = msgs.begin() + msgsPos; it != msgs.end() && y <= (int)(height / 16); ++it) {
 
 				std::string msg = *it;
-				for (int x = 0; x < msg.size() && y <= (height / 16); x += (width / 8), y++) {
+				for (int x = 0; x < msg.size() && y <= (int)(height / 16); x += (width / 8), y++) {
 					std::string msg_substr = msg.substr(x, (width / 8));
 					if(y > 0) XftDrawStringUtf8(draw, &color, font, 0, (y * 16) - 2, (const FcChar8*)msg_substr.c_str(), msg_substr.size());
 				}
